AuthService/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
## The following requirements were added by pip freeze:
django-cors-headers==4.4.0
